Remove async references
