+++ kanidmd/score/tests/common.rs
@@ -4,7 +4,7 @@ use std::sync::atomic::{AtomicU16, Ordering};
 use kanidm::audit::LogLevel;
 use kanidm::config::{Configuration, IntegrationTestConfig, ServerRole};
 use kanidm::tracing_tree;
-use kanidm_client::{asynchronous::KanidmAsyncClient, KanidmClientBuilder};
+use kanidm_client::{KanidmClient, KanidmClientBuilder};
 use score::create_server_core;
 use tokio::task;
 
@@ -21,7 +21,7 @@ fn is_free_port(port: u16) -> bool {
 }
 
 // Test external behaviours of the service.
-pub async fn setup_async_test() -> KanidmAsyncClient {
+pub async fn setup_async_test() -> KanidmClient {
     let _ = tracing_tree::test_init();
 
     let mut counter = 0;
@@ -63,7 +63,7 @@ pub async fn setup_async_test() -> KanidmAsyncClient {
     let rsclient = KanidmClientBuilder::new()
         .address(addr)
         .no_proxy()
-        .build_async()
+        .build()
         .expect("Failed to build client");
 
     rsclient
